Allow iterator to walk backwards for pos = -1
While only walking forward in the iterator, and when a request exceeds the 100m timer (`time_limit_exceeded_error`), old results could possibly be returned as the newest results.

By forcing the iterator to walk backwards from the most recent to the oldest, the most recent will always be returned and older records instead are truncated.